backend/pyproject.toml
[project]
name = "manga-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "email-validator>=2.0.0",
    "asyncpg>=0.29.0",
]

[project.scripts]
dev = "uvicorn src.main:app --reload --host 0.0.0.0 --port 8000"
